reorder code into several files
